bug: int not supported
git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/trunk/pywikipedia@2039 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6

+++ userinterfaces/tkinter_interface.py
@@ -169,7 +169,7 @@ class UI:
             # d = tkMessageBox.showinfo('title', text)
         elif urgency >= 1:
             # Save the line number before appending text
-            lineCount = int(self.logBox.index(END).split('.')[0]) - 1
+            lineCount = float(self.logBox.index(END).split('.')[0]) - 1
             self.logBox.insert(END, text)
             if colors:
                 # How many characters we already added in this line
